apps/comments_async: rm extra async comments, updating template tag, update filter display
