+++ adhocracy4/comments_async/static/comments_async/comment_box.jsx
@@ -1,3 +1,4 @@
+import $ from 'jquery'
 import React from 'react'
 import PropTypes from 'prop-types'
 import django from 'django'
@@ -25,9 +26,10 @@ export default class CommentBox extends React.Component {
     super(props)
 
     this.anchoredCommentFound = this.anchoredCommentFound.bind(this)
+    this.handleClickFilter = this.handleClickFilter.bind(this)
     this.handleClickResult = this.handleClickResult.bind(this)
-    this.handleClickSearch = this.handleClickSearch.bind(this)
     this.handleClickSorted = this.handleClickSorted.bind(this)
+    this.handleClickSearch = this.handleClickSearch.bind(this)
     this.handleEnterSearch = this.handleEnterSearch.bind(this)
     this.fetchComments = this.fetchComments.bind(this)
     this.handleCommentDelete = this.handleCommentDelete.bind(this)
@@ -38,13 +40,13 @@ export default class CommentBox extends React.Component {
     this.hideNewError = this.hideNewError.bind(this)
     this.handleHideReplyError = this.handleHideReplyError.bind(this)
     this.updateStateComment = this.updateStateComment.bind(this)
+    this.handleToggleFilters = this.handleToggleFilters.bind(this)
 
     this.state = {
       comments: [],
       nextComments: null,
       commentCount: 0,
-      showSearch: 0,
-      displayForm: false,
+      showFilters: false,
       filter: 'all',
       filterDisplay: django.gettext('all'),
       sort: 'new',
@@ -60,6 +62,7 @@ export default class CommentBox extends React.Component {
     window.addEventListener('scroll', this.handleScroll)
 
     const params = {}
+    params.ordering = this.state.sort
     if (this.state.anchoredCommentId) {
       params.commentID = this.state.anchoredCommentId
     }
@@ -76,7 +79,6 @@ export default class CommentBox extends React.Component {
                 comments: data.results,
                 nextComments: data.next,
                 commentCount: data.count,
-                showSearch: data.count,
                 anchoredCommentParentId: data.comment_parent
               }
             )
@@ -100,7 +102,6 @@ export default class CommentBox extends React.Component {
                 comments: data.results,
                 nextComments: data.next,
                 commentCount: data.count,
-                showSearch: data.count,
                 loading: false
               }
             )
@@ -132,17 +133,19 @@ export default class CommentBox extends React.Component {
         comment.displayNotification = true
         var comments = this.state.comments
         var diff = {}
+        var commentCount = this.state.commentCount
         if (typeof parentIndex !== 'undefined') {
           diff[parentIndex] = { child_comments: { $push: [comment] } }
         } else {
           diff = { $unshift: [comment] }
+          commentCount++
         }
         this.setState({
           comments: update(comments, diff),
-          commentCount: (this.state.commentCount + 1),
-          showSearch: (this.state.showSearch + 1)
+          commentCount: commentCount
         })
-
+        $('#sticky-footer').fadeIn()
+        $('body').addClass('body-sticky-footer')
         if (typeof parentIndex !== 'undefined') {
           this.updateStateComment(
             parentIndex,
@@ -159,7 +162,7 @@ export default class CommentBox extends React.Component {
         }
       })
       .fail((xhr, status, err) => {
-        var errorMessage = xhr.responseJSON.comment[0]
+        var errorMessage = (Object.values(xhr.responseJSON))[0]
         if (typeof parentIndex !== 'undefined') {
           this.updateStateComment(
             parentIndex,
@@ -198,7 +201,7 @@ export default class CommentBox extends React.Component {
         )
       })
       .fail((xhr, status, err) => {
-        var errorMessage = xhr.responseJSON.comment[0]
+        var errorMessage = Object.values(xhr.responseJSON)[0]
         this.updateStateComment(
           index,
           parentIndex,
@@ -286,32 +289,47 @@ export default class CommentBox extends React.Component {
     }
   }
 
-  handleClickSearch (e) {
+  handleToggleFilters (e) {
     e.preventDefault()
-    var search = e.currentTarget.parentElement.firstChild.value
-    this.fetchSearch(search)
+    this.setState(s => ({ showFilters: !s.showFilters }))
+  }
+
+  handleClickFilter (e) {
+    e.preventDefault()
+    var filter = e.target.id
+    this.fetchFiltered(filter)
     this.setState({
       loadingFilter: true
     })
   }
 
-  handleEnterSearch (e) {
-    var search = e.target.parentElement.firstChild.value
-    if (e.key === 'Enter') {
-      this.fetchSearch(search)
-      this.setState({
-        loadingFilter: true
-      })
+  fetchFiltered (filter) {
+    var commentCategory = filter
+    var displayFilter = this.props.commentCategoryChoices[filter]
+    if (filter === 'all') {
+      displayFilter = django.gettext('all')
+      commentCategory = ''
     }
-  }
-
-  handleClickResult (e) {
-    var result = ''
-    this.fetchSearch(result)
-    document.getElementById('search-input').value = ''
-    this.setState({
-      loadingFilter: true
+    axios.get(this.props.commentsApiUrl, {
+      params: {
+        comment_category: commentCategory,
+        ordering: this.state.sort,
+        search: this.state.search
+      }
     })
+      .then(
+        (result) => {
+          var data = result.data
+          this.setState({
+            comments: data.results,
+            nextComments: data.next,
+            commentCount: data.count,
+            filter: filter,
+            filterDisplay: displayFilter,
+            loadingFilter: false
+          })
+        }
+      )
   }
 
   handleClickSorted (e) {
@@ -349,6 +367,25 @@ export default class CommentBox extends React.Component {
       )
   }
 
+  handleClickSearch (e) {
+    e.preventDefault()
+    var search = e.currentTarget.parentElement.firstChild.value
+    this.fetchSearch(search)
+    this.setState({
+      loadingFilter: true
+    })
+  }
+
+  handleEnterSearch (e) {
+    var search = e.currentTarget.value
+    if (e.key === 'Enter') {
+      this.fetchSearch(search)
+      this.setState({
+        loadingFilter: true
+      })
+    }
+  }
+
   fetchSearch (search) {
     var commentCategory = this.state.filter
     if (commentCategory === 'all') {
@@ -373,6 +410,15 @@ export default class CommentBox extends React.Component {
           })
         }
       )
+  }
+
+  handleClickResult (e) {
+    var result = ''
+    this.fetchSearch(result)
+    document.getElementById('search-input').value = ''
+    this.setState({
+      loadingFilter: true
+    })
   }
 
   anchoredCommentFound () {
@@ -440,37 +486,81 @@ export default class CommentBox extends React.Component {
     }
   }
 
+  commentCategoryChoices () {
+    if (this.props.withCategories === true) {
+      return this.props.commentCategoryChoices
+    }
+  }
+
   render () {
     return (
       <div>
-        <div className="a4-comments__box">
-          <div className="container">
-            <CommentForm
-              subjectType={this.props.subjectType} subjectId={this.props.subjectId}
-              onCommentSubmit={this.handleCommentSubmit}
-              placeholder={django.gettext('Write contribution')}
-              rows="5" isReadOnly={this.props.isReadOnly} error={this.state.error}
-              errorMessage={this.state.errorMessage} handleErrorClick={this.hideNewError}
-              commentCategoryChoices={this.props.commentCategoryChoices}
-            />
-          </div>
-          <div className={(this.state.comments.length === 0 && this.state.loading) ? 'd-none' : 'container a4-a4-comments__nav_parent'}>
-            <nav className={this.state.showSearch === 0 ? 'd-none' : 'a4-comments__nav'}>
+        <div className="a4-comments__commentbox__form">
+          <CommentForm
+            subjectType={this.props.subjectType}
+            subjectId={this.props.subjectId}
+            onCommentSubmit={this.handleCommentSubmit}
+            placeholder={django.gettext('Write contribution')}
+            rows="5" isReadOnly={this.props.isReadOnly}
+            error={this.state.error}
+            errorMessage={this.state.errorMessage}
+            handleErrorClick={this.hideNewError}
+            commentCategoryChoices={this.commentCategoryChoices()}
+            withCategories={this.props.withCategories}
+          />
+        </div>
 
-              <div className="row flex-md-nowrap">
-                <div className="input-group a4-comments__nav__search">
-                  <input type="search" id="search-input" onKeyPress={this.handleEnterSearch} placeholder={django.gettext('Search contributions')} className="a4-comments__nav__search-input mb-0" />
-                  <button className="a4-comments__nav__search-btn btn" type="button" onClick={this.handleClickSearch}><i className="fas fa-search a4-comments__nav__search-icon" aria-label={django.gettext('Search contributions')} /></button>
+        <div className={(this.state.comments.length === 0 && this.state.loading) ? 'd-none' : 'container a4-comments__filters__parent'}>
+
+          {this.state.showFilters &&
+            <nav className="a4-comments__filters">
+
+              <div className="row flex-md-nowrap justify-content-between">
+
+                <div className="input-group a4-comments__filters__search mr-md-3">
+                  <input className="form-control a4-comments__filters__search-input mb-0" type="search" id="search-input" onKeyPress={this.handleEnterSearch} placeholder={django.gettext('Search contributions')} />
+
+                  <button className={this.state.search !== '' ? 'a4-comments__filters__search-result text-muted' : 'd-none'} type="button" onClick={this.handleClickResult}><span className="fa-stack fa-2x"><i className="far fa-circle fa-stack-2x" /><i className="fas fa-times fa-stack-1x" aria-label={django.gettext('Clear search')} /></span></button>
+
+                  <button className="input-group-append a4-comments__filters__search-btn btn btn--transparent" type="button" onClick={this.handleClickSearch}><i className="fas fa-search" aria-label={django.gettext('Search contributions')} /></button>
                 </div>
 
-                <div className="a4-comments__nav__dropdown ml-auto mr-1">
+                {this.props.withCategories &&
+                  <div className="a4-comments__filters__dropdown mr-md-3">
+                    <div className="dropdown">
+                      <button
+                        className="btn btn--transparent btn--select dropdown-toggle a4-comments__filters__btn" type="button"
+                        id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"
+                      >
+                        <span className={this.state.filter === 'all' ? 'a4-comments__filters__btn-text' : 'd-none'}>{django.gettext('display: ')}{this.state.filterDisplay}</span>
+                        <span className={this.state.filter !== 'all' ? 'a4-comments__filters__btn-text small-screen' : 'd-none'}>{this.state.filterDisplay}</span>
+
+                        <i className={this.state.filter === 'all' ? 'fa fa-caret-down' : 'fas fa-check'} aria-hidden="true" />
+                      </button>
+                      <div className="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenuButton">
+                        {this.state.filter !== 'all' &&
+                          <button className="dropdown-item" onClick={this.handleClickFilter} id="all" key="all" href="#">
+                            {django.gettext('all')}
+                          </button>}
+                        {Object.keys(this.props.commentCategoryChoices).map(objectKey => {
+                          var name = this.props.commentCategoryChoices[objectKey]
+                          if (objectKey !== this.state.filter) {
+                            return (
+                              <button className="dropdown-item" onClick={this.handleClickFilter} id={objectKey} key={objectKey} href="#">{name}</button>
+                            )
+                          }
+                        })}
+                      </div>
+                    </div>
+                  </div>}
+                <div className="a4-comments__filters__dropdown">
                   <div className="dropdown">
                     <button
-                      className="btn dropdown-toggle btn--select a4-comments__nav__btn" type="button"
+                      className="btn btn--transparent btn--select dropdown-toggle a4-comments__filters__btn" type="button"
                       id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"
                     >
-                      <span className={this.state.sort === 'new' ? 'a4-comments__nav__btn-text' : 'd-none'}>{django.gettext('sorted by: ')}{sorts[this.state.sort]}</span>
-                      <span className={this.state.sort !== 'new' ? 'a4-comments__nav__btn-text small-screen' : 'd-none'}>{sorts[this.state.sort]}</span>
+                      <span className={this.state.sort === 'new' ? 'a4-comments__filters__btn-text' : 'd-none'}>{django.gettext('sorted by: ')}{sorts[this.state.sort]}</span>
+                      <span className={this.state.sort !== 'new' ? 'a4-comments__filters__btn-text small-screen' : 'd-none'}>{sorts[this.state.sort]}</span>
                       <i className={this.state.sort === 'new' ? 'fa fa-caret-down' : 'fas fa-check'} aria-hidden="true" />
                     </button>
                     <div className="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenuButton">
@@ -490,24 +580,32 @@ export default class CommentBox extends React.Component {
                   </div>
                 </div>
               </div>
+            </nav>}
 
-              <div className="row">
-                <div className={this.state.search !== '' ? 'a4-comments__nav__text a4-comments__nav__text--wide mt-2' : 'd-none'}>
-                  <span className="a4-comments__nav__span">{this.state.commentCount + ' ' + django.ngettext('entry found for ', 'entries found for ', this.state.commentCount)}</span>
-
-                  <button className="btn btn--transparent btn--round btn--small a4-comments__nav__search-result" type="button" onClick={this.handleClickResult}>{this.state.search}<i className="fas fa-times ml-1" aria-label={django.gettext('Clear search')} /></button>
-                </div>
-
-                <div className={this.state.search === '' ? 'a4-comments__nav__text a4-comments__nav__text--wide mt-2' : 'd-none'}>
-                  {this.state.commentCount + ' ' + django.ngettext('entry', 'entries', this.state.commentCount)}
-                </div>
-              </div>
-
-            </nav>
-            <div className={this.state.loadingFilter ? 'row p-3 justify-content-center a4-comments__loading' : 'd-none'}>
-              <i className="fa fa-spinner fa-pulse" />
+          <div className="row justify-content-between">
+            <div className={this.state.search === '' ? 'a4-comments__filters__text text-muted mt-2' : 'd-none'}>
+              {this.state.commentCount + ' ' + django.ngettext('entry', 'entries', this.state.commentCount)}
             </div>
+
+            <div className={this.state.search !== '' ? 'a4-comments__filters__text text-muted mt-2' : 'd-none'}>
+              <span className="a4-comments__filters__span">{this.state.commentCount + ' ' + django.ngettext('entry found for ', 'entries found for ', this.state.commentCount)}{this.state.search}</span>
+            </div>
+
+            {!this.state.showFilters &&
+              <button className="btn btn--none text-muted a4-comments__filters__show-btn pr-0" type="button" onClick={this.handleToggleFilters}>{django.gettext('Filters')}<i className="fas fa-sliders-h ml-2" aria-label={django.gettext('Show filters')} />
+              </button>}
+            {this.state.showFilters &&
+              <button className="btn btn--none text-muted a4-comments__filters__show-btn pr-0" type="button" onClick={this.handleToggleFilters}>{django.gettext('Hide Filters')}<i className="fas fa-times ml-2" aria-label={django.gettext('Hide filters')} />
+              </button>}
           </div>
+
+          <div className={this.state.loadingFilter ? 'row p-3 justify-content-center a4-comments__loading' : 'd-none'}>
+            <i className="fa fa-spinner fa-pulse" />
+          </div>
+
+        </div>
+
+        <div className="a4-comments__box">
           <div className="a4-comments__list">
             <CommentList
               comments={this.state.comments}
@@ -517,13 +615,15 @@ export default class CommentBox extends React.Component {
               onCommentSubmit={this.handleCommentSubmit}
               onCommentModify={this.handleCommentModify}
               isReadOnly={this.props.isReadOnly}
+              commentCategoryChoices={this.commentCategoryChoices()}
               onReplyErrorClick={this.handleHideReplyError}
               onEditErrorClick={this.handleHideEditError}
+              withCategories={this.props.withCategories}
             />
           </div>
-          <div className={this.state.loading ? 'row p-3 justify-content-center a4-comments__loading' : 'd-none'}>
-            <i className="fa fa-spinner fa-pulse" />
-          </div>
+        </div>
+        <div className={this.state.loading ? 'row p-3 justify-content-center a4-comments__loading' : 'd-none'}>
+          <i className="fa fa-spinner fa-pulse" />
         </div>
       </div>
     )
