+++ adhocracy4/comments_async/static/comments_async/comment_form.jsx
@@ -104,7 +104,9 @@ export default class CommentForm extends React.Component {
             <div className="row">
               <label htmlFor="id-comment-form" className="col-6 text-muted">{this.state.commentCharCount}/4000{django.gettext(' characters')}</label>
               <div className="a4-comments__submit d-flex col-6">
-                <button type="submit" value={django.gettext('post')} className="btn a4-comments__submit-input ml-auto">{django.gettext('post')}</button>
+                {this.props.commentCategoryChoices
+                  ? <button type="submit" value={django.gettext('post')} className="btn a4-comments__submit-input ml-auto">{django.gettext('post')}</button>
+                  : <button type="submit" value={django.gettext('Comment')} className="btn a4-comments__submit-input ml-auto">{django.gettext('post')}</button>}
               </div>
             </div>
           </form>
