+++ index.js
@@ -1,8 +1,7 @@
 module.exports = {
   ratings: require('./adhocracy4/ratings/static/ratings/react_ratings.jsx'),
   comments: require('./adhocracy4/comments/static/comments/react_comments.jsx'),
-  comments_async: require('./adhocracy4/comments_async/static/comments/index.jsx'),
-  comments_async_with_categories: require('./adhocracy4/comments_async/static/comments_with_categories/index.jsx'),
+  comments_async: require('./adhocracy4/comments_async/static/comments_async/index.jsx'),
   reports: require('./adhocracy4/reports/static/reports/react_reports.jsx'),
   follows: require('./adhocracy4/follows/static/follows/react_follows.jsx'),
   polls: require('./adhocracy4/polls/static/polls/react_polls.jsx'),
